+++ scripts/docker/buildx.sh
@@ -5,20 +5,23 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 cd "$PARENT"
 
 [ "$1" == "--push" ] && push="--push" || push="--load"
+[ -z "$PYTHON_BIN" ] && PYTHON_BIN="python"
 
 ### Experimental features must be enabled and docker buildx must be installed.
 ### Run setup.sh to ensure everything is set up.
 
+mrsm_version=$($PYTHON_BIN -m meerschaum -V | sed 's/Meerschaum v//g')
+
 ./scripts/docker/update_requirements.sh
 
 docker pull "$python_image"
-for t in "${tags[@]}"; do
-  [ "$latest_alias" == "$t" ] && tag_latest="-t $image:latest" || unset tag_latest
-  docker buildx build ${push:-} \
-    --build-arg dep_group="$t" \
-    --platform "$platforms" \
-    -t "$image:$t" ${tag_latest:-} \
-    . || exit 1
-done
-# docker build --squash -t "$image" . || exit 1
+for tag in "${tags[@]}"; do
+  [ "$latest_alias" == "$tag" ] && tag_latest="-t $image:latest -t $image:$mrsm_version" || unset tag_latest
 
+  docker buildx build ${push:-} \
+    --build-arg dep_group="$tag" \
+    --platform "$platforms" \
+    -t "$image:$tag" -t "$image:$mrsm_version-$tag" ${tag_latest:-} \
+    . || exit 1
+
+done